Move group/membership code to concern
